AMB-1475: Renamed pipeline var `pr_number`
We have a pr_number variable in `project.yml` that is used on the pr-pipeline at compile-time.

In order to prevent clashes with this, the variable declared on the pipeline UI has been renamed to `action_pr_number`.

This variable is overriden via the github action triggered on a PR close event and MUST be set on the pipeline, not in the yaml, as we can't override variable values declared in the yaml files.